+++ server/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "server.js",
   "scripts": {
-    "seed": "ts-node ./seed",
+    "seed": "ts-node ./scripts/seed",
     "server": "nodemon server",
     "lint": "eslint . --ext .ts --ext .js --fix"
   },
