+++ server/server.ts
@@ -1,12 +1,12 @@
 import { ApolloServer } from 'apollo-server'
-import { connectDB } from './mongoConnection'
+import { connectDB } from './database/mongoConnection'
 import dotenv from 'dotenv'
 
 import { config } from './config'
 import RequestDataSource from './datasources/requestsDataSource'
-import { RequestsCache } from './cache'
-import { resolvers } from './resolvers'
-import { typeDefs } from './schema'
+import { RequestsCache } from './database/cache'
+import { resolvers } from './graphql/resolvers'
+import { typeDefs } from './graphql/schema'
 
 // TODO: need to make script to build(compile) prod server and to run prod server
